Update Transaksi Pembelian

+++ VPS_APP/process_unit/src/app/rincian_pesanan_pembelian_barang/rincianPesananPembelianBarang.services.js
@@ -1,5 +1,7 @@
+import { formatDate } from "../../utils/jurnalUmumUtil.js"
 import { LOGGER, LOGGER_MONITOR, logType } from "../../utils/loggerUtil.js"
 import { generatePaginationResponse } from "../../utils/paginationUtil.js"
+import { getPesananPembelianBarangByUuidService, getTanggalTransaksiTerakhirByPesananPembelianService } from "../pesanan_pembelian_barang/pesananPembelianBarang.services.js"
 import { createRincianPesananPembelianBarangRepo, deleteRincianPesananPembelianBarangByUuidRepo, getAllRincianPesananPembelianBarangRepo, getRincianPesananPembelianBarangByPesananPembelianUUIDRepo, getRincianPesananPembelianBarangByUuidRepo, updateRincianPesananPembelianBarangByUuidRepo } from "./rincianPesananPembelianBarang.repository.js"
 
 export const getAllRincianPesananPembelianBarangService = async (query, req_identity) => {
@@ -18,7 +20,7 @@ export const getAllRincianPesananPembelianBarangService = async (query, req_iden
     LOGGER(logType.INFO, "Pagination", {
         pageNumber, size, search
     }, req_identity)
-    
+
     const rincianPesananPembelianBarangs = await getAllRincianPesananPembelianBarangRepo(pageNumber, size, search, req_identity)
     return generatePaginationResponse(rincianPesananPembelianBarangs.entry, rincianPesananPembelianBarangs.count, rincianPesananPembelianBarangs.pageNumber, rincianPesananPembelianBarangs.size)
 }
@@ -46,13 +48,36 @@ export const createRincianPesananPembelianBarangService = async (rincianPesananP
     LOGGER(logType.INFO, `Start createRincianPesananPembelianBarangService`, rincianPesananPembelianBarangData, req_identity)
     rincianPesananPembelianBarangData.enabled = 1
 
+    const pesananPembelianBarang = await getPesananPembelianBarangByUuidService(rincianPesananPembelianBarangData.pesanan_pembelian_barang, req_identity)
+
+    const allowToEdit = await getTanggalTransaksiTerakhirByPesananPembelianService(pesananPembelianBarang.uuid, pesananPembelianBarang.tanggal_pesanan_pembelian_barang, req_identity);
+
+    if (!allowToEdit.allow) {
+        throw Error(JSON.stringify({
+            message: "Tidak Bisa Diedit Karena Tanggal Terakhir Transaksi Adalah " + formatDate(allowToEdit.tanggal_minimum),
+            field: "error"
+        }))
+    }
+
     const rincianPesananPembelianBarang = await createRincianPesananPembelianBarangRepo(rincianPesananPembelianBarangData, req_identity)
     return rincianPesananPembelianBarang
 }
 
 export const deleteRincianPesananPembelianBarangByUuidService = async (uuid, req_identity) => {
     LOGGER(logType.INFO, `Start deleteRincianPesananPembelianBarangByUuidService [${uuid}]`, null, req_identity)
-    await getRincianPesananPembelianBarangByUuidService(uuid, req_identity)
+    const beforeData = await getRincianPesananPembelianBarangByUuidService(uuid, req_identity)
+
+    const pesananPembelianBarang = await getPesananPembelianBarangByUuidService(beforeData.pesanan_pembelian_barang, req_identity)
+
+    const allowToEdit = await getTanggalTransaksiTerakhirByPesananPembelianService(pesananPembelianBarang.uuid, pesananPembelianBarang.tanggal_pesanan_pembelian_barang, req_identity);
+
+    if (!allowToEdit.allow) {
+        throw Error(JSON.stringify({
+            message: "Tidak Bisa Diedit Karena Tanggal Terakhir Transaksi Adalah " + formatDate(allowToEdit.tanggal_minimum),
+            field: "error"
+        }))
+    }
+
     await deleteRincianPesananPembelianBarangByUuidRepo(uuid, req_identity)
     return true
 }
@@ -60,6 +85,18 @@ export const deleteRincianPesananPembelianBarangByUuidService = async (uuid, req
 export const updateRincianPesananPembelianBarangByUuidService = async (uuid, rincianPesananPembelianBarangData, req_identity, req_original_url, req_method) => {
     LOGGER(logType.INFO, `Start updateRincianPesananPembelianBarangByUuidService [${uuid}]`, rincianPesananPembelianBarangData, req_identity)
     const beforeData = await getRincianPesananPembelianBarangByUuidService(uuid, req_identity)
+
+    const pesananPembelianBarang = await getPesananPembelianBarangByUuidService(beforeData.pesanan_pembelian_barang, req_identity)
+
+    const allowToEdit = await getTanggalTransaksiTerakhirByPesananPembelianService(pesananPembelianBarang.uuid, pesananPembelianBarang.tanggal_pesanan_pembelian_barang, req_identity);
+
+    if (!allowToEdit.allow) {
+        throw Error(JSON.stringify({
+            message: "Tidak Bisa Diedit Karena Tanggal Terakhir Transaksi Adalah " + formatDate(allowToEdit.tanggal_minimum),
+            field: "error"
+        }))
+    }
+
     const rincianPesananPembelianBarang = await updateRincianPesananPembelianBarangByUuidRepo(uuid, rincianPesananPembelianBarangData, req_identity)
 
     LOGGER_MONITOR(req_original_url, req_method, {
